learn Protected routing
