+++ src/App.js
@@ -1,11 +1,12 @@
 import "./App.css";
-import { Header } from "./components/Header";
-import { Footer } from "./components/Footer";
+import { Header } from "./components/UI/Header";
+// import { Footer } from "./components/Footer";
 import Movies from "./components/movies";
 import { Login } from "./components/Authentication/Login";
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 import { SignUp } from "./components/Authentication/SignUp/SignUp";
-import Contex from "./components/Contex/RootContex";
+import Contex from "./Contex/RootContex";
+import ProtecedRoutes from "./ProtectedRoutes";
 
 function App() {
   return (
@@ -13,9 +14,9 @@ function App() {
     <Router>
       <Header />
       <Switch>
-        <Route exact path="/">
+        <ProtecedRoutes exact path="/">
           <Movies />
-        </Route>
+        </ProtecedRoutes>
         <Route exact path="/login">
           <Login />
         </Route>
